new file:   Chaitanya/kafka/__init__.py 	new file:   Chaitanya/kafka/commands/__init__.py 	new file:   Chaitanya/kafka/commands/v100.py 	new file:   Chaitanya/kafka/exporter/__init__.py 	new file:   Chaitanya/kafka/exporter/kafkaexporter.py 	modified:   Chaitanya/postfix/commands/v2113.py 	modified:   Chaitanya/postfix/exporter/postfix_exporter.py 	modified:   Chaitanya/redis/exporter/redisexporter.py

+++ Chaitanya/postfix/commands/v2113.py
@@ -33,7 +33,7 @@ postfix_metrices=[
     'command': "pqshell --summary |awk '/Total queue size/ {print $4}'",
     'desc': 'Size of the mail queue',
     'type': 'summary',
-    'data_type': 'float'
+    'data_type': 'integer'
     }
 
 ]
